A huge update to modify the grid class structrure of himan. This is one of the biggest revamps in himan lib core functionality, and should therefore be tested fully before put into operational use.
The primary change introduced is how different grids are used
in himan. A little bit of history: in the very beginning there
was only class 'grid', which contained all the necessary information
needed for himan. For example grid coordinates, number of gridpoints
etc. The supported projections were latlon, rotlatlon and polster.

Then the concept of irregular grids was introduced, and here irregular
grid actually means only a collection of distinct points. The original
grid class was made into a parent class, and two new classes were 
created: regular_grid (= the original grid) and irregular_grid. Everything
worked well with this setup for a (relatively) long time. Then reduced
gaussian grid came along. This grid type is irregular but not just a 
collection of points, and it behaves like a regular grid. The old class
structure just didn't work anymore. 

This commit splits each projection (hence to be known as grid type) to
its separate class, which inherits from grid base class. This split
will also be made to geom-table in radon database. The regular/irregular
grid division is now a property of the class, known as grid class. The 
old irregular_grid class has been renamed to point_list. At the same time
all interpolation -related (except CUDA) functions have been moved to
a new namespace interpolation. Support for to/from reduced gaussian grid
has been implemented.

Brief list of changes:
* regular_grid and irregular_grid have been split to latitude_longitude_grid,
  rotated_latitude_longitude_grid, stereographic_grid and point_list
* New class and functionality reduced_gaussian_grid introduced
* Interpolation functions moved to interpolation-namespace in himan-lib
* Functions defined by the parent grid-class have been extended a bit,
  to avoid repeated dynamic_casts in the deriving classes
* Unit tests have been created for the new grid class hierarchy

What's still missing:
* Full testing, ie running himan-test suite
* Testing with GPU: all changes were made on a non-GPU based system so 
  that code has NOT been tested yet
* Database changes
* Unit tests for interpolation functions

+++ himan-lib/source/param.cpp
@@ -384,7 +384,7 @@ ostream& param::Write(ostream& file) const
 	file << "__itsGribDiscipline__ " << itsGribDiscipline << endl;
 	file << "__itsGribTableVersion__ " << itsGribTableVersion << endl;
 	file << "__itsIndicatorOfParameter__ " << itsGribIndicatorOfParameter << endl;
-	file << "__itsUnit__ " << itsUnit << endl;
+	file << "__itsUnit__ " << static_cast<int> (itsUnit) << endl;
 	file << "__itsVersion__ " << itsVersion << endl;
 	file << "__itsInterpolationMethod__ " << HPInterpolationMethodToString.at(itsInterpolationMethod) << endl;
